+++ WeaselServer/WeaselServerApp.h
@@ -9,8 +9,6 @@
 #include <filesystem>
 #include <functional>
 #include <memory>
-#include <winsparkle.h>
-
 #include "WeaselTrayIcon.h"
 
 namespace fs = std::filesystem;
@@ -33,21 +31,7 @@ class WeaselServerApp {
                                     SW_SHOWNORMAL) > 32;
   }
 
-  static bool check_update() {
-    // when checked manually, show testing versions too
-    std::string feed_url = GetCustomResource("ManualUpdateFeedURL", "APPCAST");
-    std::wstring channel{};
-    auto ret = RegGetStringValue(HKEY_CURRENT_USER, L"Software\\Rime\\Weasel",
-                                 L"UpdateChannel", channel);
-    if (!ret && channel == L"testing") {
-      feed_url = GetCustomResource("TestingManualUpdateFeedURL", "APPCAST");
-    }
-    if (!feed_url.empty()) {
-      win_sparkle_set_appcast_url(feed_url.c_str());
-    }
-    win_sparkle_check_update_with_ui();
-    return true;
-  }
+  static bool check_update();
 
   static fs::path install_dir() {
     WCHAR exe_path[MAX_PATH] = {0};